This patch provides initial experimental support for Date/TZ for MSVC and MingW platforms. Currently VS2015 and C++14 is required for Win32. Linux should work with C++11. Use of this library with Win32 requires that TimeZoneMappings.csv exist in the install folder where the rest of the timezone database is installed.
It's a proof of concept at this stage in quite how the timezone mappings element will be integrated
into the date library so it's subject to change but should settle down soon.
Note that on Linux platforms all code relating to the TIME_MAPPINGS macro should be compiled out by default.
This macro may also disappear in a later revision.