Don't delete the gz/tar files on error. This change makes semantics the same Windows as Linux.
It's arguable if files should be left on error but the semantics should be the same between OS's which is what this patch is about.
It does offer a small chance for a user to investigate the files if something does go wrong.
e.g. If the disk filled up or someone removed the zip program. Then they could try again manually to recover.