Add a configuration header
This makes it easier to install and use the tz library correctly,
since the correct defines, which were used when tz was compiled
will be set by the configuration header which is automatically
included by tz.h

This follows the pattern set by common open source libraries like
libpng and freetype.

This commit also fixes a bug where the define USE_AUTOLOAD changed
name to AUTO_DOWNLOAD but the cmake was not updated correspondingly.

This was caught because we added

at the top of the tz.h header after we include the configuration
header.