build: bump minimum cmake version to 3.8 to make sure `cxx_std_11` exists.
I increase the minimum CMake version to 3.8 because that's when the cxx_std_11 compile feature was added. Realistically this requirement could be increased to at least 3.10 since 3.10 is what Ubuntu 18 ships with and it's unlikely that any users are locked to an older CMake version than that.

Text copied from https://github.com/fmtlib/fmt/pull/3205